refactor(tutti-au-host): Phase 4 — replace hand-rolled Apple FFI with binding crates
Swap ~830 LOC of hand-declared CoreAudio/CoreFoundation/Cocoa FFI for the
workspace binding crates (coreaudio-sys, core-foundation 0.10, objc2 /
objc2-foundation). Net −259 LOC; types.rs 721→~250. AU-only, macOS-only.
No behavior change — this is a working-code → idiomatic-code swap, done under a
strict do-no-harm mandate with the credited-correct behaviors test-pinned.

- types.rs: all AudioToolbox extern "C" fns + #[repr(C)] structs (ASBD,
  AudioBufferList, AudioTimeStamp, AudioComponentDescription, AudioUnitCocoaViewInfo,
  AudioUnitParameterInfo, AURenderCallbackStruct, NSPoint/Size/Rect + manual Encode)
  now come from coreaudio-sys; SCREAMING_CASE consts alias coreaudio_sys::k… so
  values are provably byte-identical. Deleted the blanket
  #![allow(non_upper_case_globals, non_camel_case_types, dead_code)].
- cf.rs: CfString/CfUrl/CfPlist wrap core-foundation TCFType (CFString/CFURL/
  CFPropertyList); the cf_owned! macro and all manual CFRetain/CFRelease gone.
  CfData deleted.
- editor/cocoa.rs: the raw objc_msgSend trampolines + AnyClass::get(c"NSBundle")
  replaced by objc2_foundation NSBundle/NSString/NSSize + objc2 msg_send!. The two
  genuinely-dynamic dispatches (bundleWithURL:, uiViewForAudioUnit:withSize: on the
  plugin-provided factory class) stay idiomatic objc2 — not hand-rolled FFI.
- ffi.rs: fix AudioUnitGetPropertyInfo outWritable — was &mut i32, is *mut Boolean(u8).

Preserved + test-pinned: ASBD non-interleaved float32 flag math (new
test_asbd_float32_exact_fields asserts every field for mono+stereo, incl.
mBytesPerFrame=4 not N*4); CF Create/Get ownership balancing (Create-rule for
ClassInfo/CocoaUI/CopyName, Get-rule for borrowed); the Phase-1 render-callback-once
+ pinned-scratch ref_con (test_render_callback_installed_once still 1 install across
16 blocks); StreamFormat re-probe; single main-thread editor teardown; Phase-2
send_midi + real param flags. AudioBufferList FAM layout verified identical.

27 AU tests + the no-alloc RT test green; warning- and clippy-clean.

+++ crates/plugin/formats/tutti-au-host/src/ffi.rs
@@ -85,7 +85,8 @@ pub(crate) unsafe fn property_size(
     element: u32,
 ) -> Result<u32> {
     let mut size: u32 = 0;
-    let mut writable: i32 = 0;
+    // `outWritable` is `*mut Boolean` (a C `unsigned char`) in AudioToolbox.
+    let mut writable: u8 = 0;
     check(
         "AudioUnitGetPropertyInfo",
         AudioUnitGetPropertyInfo(unit, id, scope, element, &mut size, &mut writable),
